enable javascript hot reload
