+++ index.html
@@ -6,7 +6,7 @@
   </head>
   <body>
     <h1>Show files in folder</h1>
-    <form onsubmit="return renderer.attachElement(this.path.value)">
+    <form onsubmit="renderer.attachElement(this.path.value);return false">
       Path:<br>
       <input type="text" name="path"><br>
       <input type="submit" value="Submit">
@@ -26,8 +26,11 @@
   </body>
 
   <script>
-    const renderer = require('./renderer.js');
-    const dicomParser = require('dicom-parser');
-    const dicom = require('./dicom.js');
+    // require("./renderer.js")
+    (function() {
+      const script = document.createElement('script');
+      script.src = (process.env.HOT) ? 'http://localhost:8080/build/app.bundle.js' : 'build/app.bundle.js';
+      document.write(script.outerHTML);
+    }());
   </script>
 </html>
